📦 Update Hey-pipi and Hey-pee-dar Models and Pinout Image
This commit updates binary files for the "Hey-pipi" and "Hey-pee-dar" models, as well as the "pinout.jpg" image in the Raspberry Pi Zero 2 W client directory. These changes may include improvements or bug fixes to the models and updates to the image, which could enhance functionality or provide more accurate visual information for hardware connections.